GCI-142 Java 5 API optimalizations (for loop replaced by foreach, contains() instead of indexOf(), unnecessary unboxing)

+++ api/src/main/java/org/openmrs/Form.java
@@ -148,8 +148,8 @@ public class Form extends BaseChangeableOpenmrsMetadata {
 				
 				for (FormField ff : fieldSet) {
 					if (ff.getFieldNumber() != null) {
-						if (ff.getFieldNumber().intValue() < fieldNum || fieldNum == 0) {
-							fieldNum = ff.getFieldNumber().intValue();
+						if (ff.getFieldNumber() < fieldNum || fieldNum == 0) {
+							fieldNum = ff.getFieldNumber();
 							next = ff;
 						}
 					} else {
